// frontend/src/workbench/extensions/manager/types/generatedManagerTypes.ts
export interface components {
  schemas: {
// ...
    ManagerPackInstalled: {
      /** @description The version of the pack that is installed (Git commit hash or semantic version) */
      ver: string
      /** @description The name of the pack if installed from the registry */
      cnr_id?: string | null
      /** @description The name of the pack if installed from github (author/repo-name format) */
      aux_id?: string | null
      /** @description Whether the pack is enabled */
      enabled: boolean
    }
    /**
     * @description Version selection for pack installation
     * @enum {string}
     */
    SelectedVersion: 'latest' | 'nightly'
    /**
     * @description Channel for pack sources
     * @enum {string}
     */
    ManagerChannel:
      | 'default'
      | 'recent'
      | 'legacy'
      | 'forked'
      | 'dev'
      | 'tutorial'
    /**
     * @description Source for pack information
     * @enum {string}
     */
    ManagerDatabaseSource: 'remote' | 'local' | 'cache'
// ...
    InstalledPacksResponse: {
      [key: string]: components['schemas']['ManagerPackInstalled']
    }
// ...
  }
// ...
}